update old option & chang directory
