--- integrations/EasyEDA/Makefile
@@ -1,5 +1,0 @@
-build:
-	g++ -O2 freeroutingp.cpp -o freeroutingp
-clean:
-	-@rm freeroutingp
-	-@rm freeroutingp.exe
